Fixed warnings during doc build
Several build warnings where due to lack of titles or placeholder files for past changelogs that were never filled and probably won't ever be. I delete unecessary files and fixed the missing titles to fix the warnings. Some of these files still need to be filled up with info, but this won't be fixed right now.